Update natverse packages

This will check to see if all natverse packages (and optionally (if recursive = TRUE), their dependencies ) are up-to-date, and will provide the command to perform the installation in one go! Adapted and modified from the following sources : update function in 'tidyverse' package, github_update function in 'dtupdate' package

natverse_update(recursive = FALSE, source = c("CRAN", "GITHUB"))

Arguments

recursive
If TRUE, will also check all dependencies of natverse packages.

source
set the source of updates to 'CRAN' or 'GITHUB' natverse packages.

Value

A tibble detailing dependencies and their status.

Examples

natverse_update()

#> 
#> All natverse dependencies from CRAN are up-to-date, see details below:
#> 
#> package    cran    local   status            
#> ---------  ------  ------  ------------      
#> cli        1.1.0   1.1.0   ✔ 
#> crayon     1.3.4   1.3.4   ✔ 
#> dplyr      0.8.3   0.8.3   ✔ 
#> magrittr   1.5     1.5     ✔ 
#> pbapply    1.4.1   1.4.1   ✔ 
#> purrr      0.3.2   0.3.2   ✔ 
#> stringr    1.4.0   1.4.0   ✔ 
#> tibble     2.1.3   2.1.3   ✔