Check status of natverse packages and (optionally) update if necessary

This will check to see if all natverse packages (and by default, when recursive = TRUE), their dependencies ) are up-to-date, and will update them (optionally, if update = TRUE) if they are missing or are not up-to-date!

natverse_update(
  update = FALSE,
  install_missing = update,
  recursive = TRUE,
  upgrade = c("default", "ask", "always", "never"),
  ...
)

Arguments

update
If TRUE, will actually update the packages rather than just reporting the status of dependent packages.

install_missing
If TRUE, will install any missing packages.

recursive
If TRUE, will also check all dependencies of natverse packages.

upgrade
Whether or not to ask you about updating the dependencies of natverse packages. The 'default' value will ask in interactive session. Set to 'always' to install all dependencies without prompts. See update_packages for further details.

...
extra arguments to pass to update_packages, install_cran, install_github.

Examples

if (FALSE) {
# check status of natverse and its dependencies
natverse_update()

# ... actually install any updates
natverse_update(update=TRUE)

# update any indirect dependencies without requesting your confirmation
natverse_update(update=TRUE, upgrade='always')
}